+++ Source/fuentes/control.h
@@ -1,2 +1,2 @@
 void cargar_buffer_single(char dato);
-int cambiar_pin();
+short int cambiar_pin();
